working save and query(with empId) with testCases
What changes you have made?
all testCases are implemented to all functions and test cases for
mocha organized in good manner

Why this change is necessary?
to test functions and understand testing of mocha on file by file

What is the impact of the change?
to ensure that all functions working properly